requirements.txt
fastapi
uvicorn
pydantic
motor
pymongo
langchain
langchain-google-genai
python-dotenv